backend delete works, refactored DB to use int for ID, DUH

+++ NewToDoApp/NewToDoApp/Migrations/ApplicationDbContextModelSnapshot.cs
@@ -222,11 +222,11 @@ namespace NewToDoApp.Migrations
 
             modelBuilder.Entity("NewToDoApp.Models.TodoItem", b =>
                 {
-                    b.Property<long>("Id")
+                    b.Property<int>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("bigint");
+                        .HasColumnType("int");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("Id"));
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"));
 
                     b.Property<bool>("IsComplete")
                         .HasColumnType("bit");
